# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Vehicle {
  id: number;
  vehicleNo: string;
  vehicleName: string;
  active: boolean;
}

export interface VehicleInput {
  /** @minLength 1 */
  vehicleNo: string;
  /** @minLength 1 */
  vehicleName: string;
}

export interface VehicleUpdate {
  /** @minLength 1 */
  vehicleNo?: string;
  /** @minLength 1 */
  vehicleName?: string;
  active?: boolean;
}

export interface DieselRecord {
  id: number;
  date: string;
  vehicleId: number;
  vehicleNo: string;
  vehicleName: string;
  openingBalance: number;
  /** @nullable */
  reading: number | null;
  dieselIssued: number;
  dieselPurchased: number;
  closingBalance: number;
  operator: string;
  issuer: string;
  createdAt: string;
}

export interface DieselRecordInput {
  date: string;
  vehicleId: number;
  /** @nullable */
  reading?: number | null;
  /** @minimum 0 */
  dieselIssued: number;
  /** @minimum 0 */
  dieselPurchased: number;
  /** @minLength 1 */
  operator: string;
  /** @minLength 1 */
  issuer: string;
}

export interface DieselSummary {
  date: string;
  openingBalance: number;
  totalIssued: number;
  totalPurchased: number;
  closingBalance: number;
  recordCount: number;
  recentRecords: DieselRecord[];
}

export type GetDieselSummaryParams = {
date: string;
};

export type ListDieselRecordsParams = {
date: string;
};