# lib/api-zod/src/generated/types/dieselRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DieselRecord {
  id: number;
  date: Date;
  vehicleId: number;
  vehicleNo: string;
  vehicleName: string;
  openingBalance: number;
  /** @nullable */
  reading: number | null;
  dieselIssued: number;
  dieselPurchased: number;
  closingBalance: number;
  operator: string;
  issuer: string;
  createdAt: Date;
}