# lib/api-zod/src/generated/types/dieselRecordInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DieselRecordInput {
  date: Date;
  vehicleId: number;
  /** @nullable */
  reading?: number | null;
  /** @minimum 0 */
  dieselIssued: number;
  /** @minimum 0 */
  dieselPurchased: number;
  /** @minLength 1 */
  operator: string;
  /** @minLength 1 */
  issuer: string;
}