# lib/api-zod/src/generated/types/dieselSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DieselRecord } from './dieselRecord';

export interface DieselSummary {
  date: Date;
  openingBalance: number;
  totalIssued: number;
  totalPurchased: number;
  closingBalance: number;
  recordCount: number;
  recentRecords: DieselRecord[];
}