# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleInput {
  /** @minLength 1 */
  vehicleNo: string;
  /** @minLength 1 */
  vehicleName: string;
}